update the router and make homehandler.js #9
